Migrated primary feed controller methods to new model.
