+++ app/controllers/definitions/feed/entries/index.js
@@ -1,4 +1,4 @@
-const perms = require("../../../config/permissions");
+const perms = require("../../../../config/permissions");
 
 const {
     CommentEntryType,
