+++ app/controllers/feed.js
@@ -1,0 +1,3 @@
+const FeedController = require("./definitions/feed");
+
+module.exports = new FeedController();
